requirements.txt
requests
beautifulsoup4
sslyze
Flask
jinja2
fpdf2
python-whois
dnspython
colorama
